requirements.txt
alembic==1.8.1
anaconda-navigator==2.1.4
appdirs==1.4.4
atomicwrites==1.4.0
backports.weakref==1.0.post1
bkcharts==0.2
black==19.10b0
bravado==10.4.1
bravado-core==5.17.0
brotlipy==0.7.0
Bybit==0.2.12
bybit-ws==0.1.8
certifi==2021.10.8
clyent==1.2.2
comtypes==1.1.10
conda==4.12.0
conda-build==3.21.8
conda-verify==3.4.2
constantly==15.1.0
cssselect==1.1.0
cytoolz==0.11.0
daal4py==2021.5.0
datashape==0.5.4
distlib==0.3.6
et-xmlfile==1.1.0
Flask-Login==0.6.2
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
fonttools==4.25.0
fqdn==1.5.1
import-ipynb==0.1.4
inflection==0.5.1
isoduration==20.11.0
jsonpointer==2.3
jsonref==0.2
llvmlite==0.38.0
Mako==1.2.3
mccabe==0.6.1
mkl-fft==1.3.1
mkl-service==2.4.0
monotonic==1.6
mpmath==1.2.1
msgpack-python==0.5.6
munkres==1.1.4
mypy-extensions==0.4.3
navigator-updater==0.2.1
pathspec==0.7.0
patsy==0.5.2
pep8==1.7.1
Pillow==9.0.1
protobuf==3.19.1
pyasn1-modules==0.2.8
pybit==2.3.0
pycosat==0.6.3
pycurl==7.44.1
PyDispatcher==2.0.5
pyls-spyder==0.4.0
pyreadline==2.1
pytest==7.1.1
python-lsp-jsonrpc==1.0.0
python-lsp-server==1.2.4
pytz==2021.3
pywin32==302
PyYAML==6.0
queuelib==1.5.0
rfc3339-validator==0.1.4
scikit-learn-intelex==2021.20220215.102710
simplejson==3.17.6
sip==4.19.13
statsmodels==0.13.2
swagger-spec-validator==2.7.4
tables==3.6.1
tabulate==0.8.9
TBB==0.2
uri-template==1.2.0
urllib3==1.24.2
virtualenv==20.16.5
webcolors==1.12
webencodings==0.5.1
websocket-client==0.53.0
websockets==10.3
win-unicode-console==0.5
wincertstore==0.2
xlwings==0.24.9
zict==2.0.0
